+++ ui/src/dashboards/containers/DashboardPage.js
@@ -8,7 +8,6 @@ import _ from 'lodash'
 
 import {isUserAuthorized, EDITOR_ROLE} from 'src/auth/Authorized'
 
-import OverlayTechnologies from 'shared/components/OverlayTechnologies'
 import CellEditorOverlay from 'src/dashboards/components/CellEditorOverlay'
 import DashboardHeader from 'src/dashboards/components/DashboardHeader'
 import Dashboard from 'src/dashboards/components/Dashboard'
@@ -28,6 +27,7 @@ import {
   showCellEditorOverlay,
   hideCellEditorOverlay,
 } from 'src/dashboards/actions/cellEditorOverlay'
+import {showOverlay} from 'src/shared/actions/overlayTechnology'
 
 import {dismissEditingAnnotation} from 'src/shared/actions/annotations'
 
@@ -36,10 +36,16 @@ import {
   templateControlBarVisibilityToggled as templateControlBarVisibilityToggledAction,
 } from 'shared/actions/app'
 import {presentationButtonDispatcher} from 'shared/dispatchers'
-import {interval, DASHBOARD_LAYOUT_ROW_HEIGHT} from 'shared/constants'
+import {
+  interval,
+  DASHBOARD_LAYOUT_ROW_HEIGHT,
+  TEMP_VAR_DASHBOARD_TIME,
+  TEMP_VAR_UPPER_DASHBOARD_TIME,
+} from 'shared/constants'
 import {notifyDashboardNotFound} from 'shared/copy/notifications'
 import {colorsStringSchema, colorsNumberSchema} from 'shared/schemas'
 import {ErrorHandling} from 'src/shared/decorators/errors'
+import {OverlayContext} from 'src/shared/components/OverlayTechnology'
 
 const FORMAT_INFLUXQL = 'influxql'
 const defaultTimeRange = {
@@ -57,7 +63,6 @@ class DashboardPage extends Component {
     this.state = {
       isEditMode: false,
       selectedCell: null,
-      isTemplating: false,
       zoomedTimeRange: {zoomedLower: null, zoomedUpper: null},
       scrollTop: 0,
       windowHeight: window.innerHeight,
@@ -149,16 +154,28 @@ class DashboardPage extends Component {
   }
 
   handleOpenTemplateManager = () => {
-    this.setState({isTemplating: true})
-  }
-
-  handleCloseTemplateManager = isEdited => () => {
-    if (
-      !isEdited ||
-      (isEdited && confirm('Do you want to close without saving?')) // eslint-disable-line no-alert
-    ) {
-      this.setState({isTemplating: false})
+    const {handleShowOverlay, dashboard, source} = this.props
+    const options = {
+      dismissOnClickOutside: false,
+      dismissOnEscape: false,
     }
+
+    handleShowOverlay(
+      <OverlayContext.Consumer>
+        {({onDismissOverlay}) => {
+          return (
+            <TemplateVariableManager
+              source={source}
+              templates={dashboard.templates}
+              onDismissOverlay={onDismissOverlay}
+              onRunQueryFailure={this.handleRunQueryFailure}
+              onEditTemplateVariables={this.handleEditTemplateVariables}
+            />
+          )
+        }}
+      </OverlayContext.Consumer>,
+      options
+    )
   }
 
   handleSaveEditedCell = newCell => {
@@ -321,7 +338,7 @@ class DashboardPage extends Component {
 
     const dashboardTime = {
       id: 'dashtime',
-      tempVar: ':dashboardTime:',
+      tempVar: TEMP_VAR_DASHBOARD_TIME,
       type: lowerType,
       values: [
         {
@@ -334,7 +351,7 @@ class DashboardPage extends Component {
 
     const upperDashboardTime = {
       id: 'upperdashtime',
-      tempVar: ':upperDashboardTime:',
+      tempVar: TEMP_VAR_UPPER_DASHBOARD_TIME,
       type: upperType,
       values: [
         {
@@ -357,7 +374,7 @@ class DashboardPage extends Component {
       templatesIncludingDashTime = []
     }
 
-    const {isEditMode, isTemplating} = this.state
+    const {isEditMode} = this.state
 
     const names = dashboards.map(d => ({
       name: d.name,
@@ -365,17 +382,6 @@ class DashboardPage extends Component {
     }))
     return (
       <div className="page dashboard-page">
-        {isTemplating ? (
-          <OverlayTechnologies>
-            <TemplateVariableManager
-              source={source}
-              templates={dashboard.templates}
-              onClose={this.handleCloseTemplateManager}
-              onRunQueryFailure={this.handleRunQueryFailure}
-              onEditTemplateVariables={this.handleEditTemplateVariables}
-            />
-          </OverlayTechnologies>
-        ) : null}
         {selectedCell ? (
           <CellEditorOverlay
             source={source}
@@ -534,6 +540,7 @@ DashboardPage.propTypes = {
   thresholdsListColors: colorsNumberSchema.isRequired,
   gaugeColors: colorsNumberSchema.isRequired,
   lineColors: colorsStringSchema.isRequired,
+  handleShowOverlay: func.isRequired,
 }
 
 const mapStateToProps = (state, {params: {dashboardID}}) => {
@@ -613,6 +620,7 @@ const mapDispatchToProps = dispatch => ({
     dismissEditingAnnotation,
     dispatch
   ),
+  handleShowOverlay: bindActionCreators(showOverlay, dispatch),
 })
 
 export default connect(mapStateToProps, mapDispatchToProps)(