+++ ui/src/data_explorer/components/GroupByTimeDropdown.tsx
@@ -1,24 +1,40 @@
-import React from 'react'
-import PropTypes from 'prop-types'
+import React, {SFC} from 'react'
 import {withRouter} from 'react-router'
+import {Location} from 'history'
 
 import groupByTimeOptions from 'src/data_explorer/data/groupByTimes'
 
-import Dropdown from 'shared/components/Dropdown'
+import Dropdown from 'src/shared/components/Dropdown'
 
-import {AUTO_GROUP_BY} from 'shared/constants'
+import {AUTO_GROUP_BY} from 'src/shared/constants'
+import {GroupBy} from 'src/types'
 
-const isInRuleBuilder = pathname => pathname.includes('alert-rules')
+interface GroupByTimeOption {
+  defaultTimeBound: string
+  seconds: number
+  menuOption: string
+}
 
-const getOptions = pathname =>
+interface Props {
+  location?: Location
+  selected: string
+  onChooseGroupByTime: (groupBy: GroupBy) => void
+  isDisabled: boolean
+}
+
+const isInRuleBuilder = (pathname: string): boolean =>
+  pathname.includes('alert-rules')
+
+const getOptions = (pathname: string): GroupByTimeOption[] =>
   isInRuleBuilder(pathname)
     ? groupByTimeOptions.filter(({menuOption}) => menuOption !== AUTO_GROUP_BY)
     : groupByTimeOptions
 
-const GroupByTimeDropdown = ({
+const GroupByTimeDropdown: SFC<Props> = ({
   selected,
   onChooseGroupByTime,
   location: {pathname},
+  isDisabled,
 }) => (
   <div className="group-by-time">
     <label className="group-by-time--label">Group by:</label>
@@ -32,18 +48,9 @@ const GroupByTimeDropdown = ({
       }))}
       onChoose={onChooseGroupByTime}
       selected={selected || 'Time'}
+      disabled={isDisabled}
     />
   </div>
 )
-
-const {func, string, shape} = PropTypes
-
-GroupByTimeDropdown.propTypes = {
-  location: shape({
-    pathname: string.isRequired,
-  }).isRequired,
-  selected: string,
-  onChooseGroupByTime: func.isRequired,
-}
 
 export default withRouter(GroupByTimeDropdown)