+++ ui/src/shared/components/QueryOptions.tsx
@@ -1,10 +1,23 @@
-import React from 'react'
-import PropTypes from 'prop-types'
+import React, {SFC} from 'react'
+
+import {GroupBy, TimeShift} from 'src/types'
+
 import GroupByTimeDropdown from 'src/data_explorer/components/GroupByTimeDropdown'
 import TimeShiftDropdown from 'src/shared/components/TimeShiftDropdown'
-import FillQuery from 'shared/components/FillQuery'
+import FillQuery from 'src/shared/components/FillQuery'
 
-const QueryOptions = ({
+interface Props {
+  fill: string
+  onFill: (fill: string) => void
+  groupBy: GroupBy
+  shift: TimeShift
+  onGroupByTime: (groupBy: GroupBy) => void
+  isKapacitorRule: boolean
+  onTimeShift: (shift: TimeShift) => void
+  isDisabled: boolean
+}
+
+const QueryOptions: SFC<Props> = ({
   fill,
   shift,
   onFill,
@@ -12,36 +25,25 @@ const QueryOptions = ({
   onTimeShift,
   onGroupByTime,
   isKapacitorRule,
+  isDisabled,
 }) => (
   <div className="query-builder--groupby-fill-container">
     <GroupByTimeDropdown
       selected={groupBy.time}
       onChooseGroupByTime={onGroupByTime}
+      isDisabled={isDisabled}
     />
     {isKapacitorRule ? null : (
       <TimeShiftDropdown
         selected={shift && shift.label}
         onChooseTimeShift={onTimeShift}
+        isDisabled={isDisabled}
       />
     )}
-    {isKapacitorRule ? null : <FillQuery value={fill} onChooseFill={onFill} />}
+    {isKapacitorRule ? null : (
+      <FillQuery value={fill} onChooseFill={onFill} isDisabled={isDisabled} />
+    )}
   </div>
 )
-
-const {bool, func, shape, string} = PropTypes
-
-QueryOptions.propTypes = {
-  fill: string,
-  onFill: func.isRequired,
-  groupBy: shape({
-    time: string,
-  }).isRequired,
-  shift: shape({
-    label: string,
-  }),
-  onGroupByTime: func.isRequired,
-  isKapacitorRule: bool.isRequired,
-  onTimeShift: func.isRequired,
-}
 
 export default QueryOptions