+++ ui/src/types/index.ts
@@ -1,8 +1,17 @@
 import {AuthLinks, Organization, Role, User, Me} from './auth'
-import {Query, QueryConfig, TimeRange} from './query'
+import {Template, Cell, CellQuery, Legend} from './dashboard'
+import {
+  GroupBy,
+  QueryConfig,
+  Status,
+  TimeRange,
+  TimeShift,
+  Field,
+} from './query'
 import {AlertRule, Kapacitor, Task} from './kapacitor'
-import {Source} from './sources'
+import {Source, SourceLinks} from './sources'
 import {DropdownAction, DropdownItem} from './shared'
+import {Notification} from 'src/kapacitor/components/AlertOutputs'
 
 export {
   Me,
@@ -10,13 +19,22 @@ export {
   Role,
   User,
   Organization,
+  Template,
+  Cell,
+  CellQuery,
+  Legend,
+  Status,
+  QueryConfig,
+  TimeShift,
+  Field,
+  GroupBy,
   AlertRule,
   Kapacitor,
-  Query,
-  QueryConfig,
   Source,
+  SourceLinks,
   DropdownAction,
   DropdownItem,
   TimeRange,
   Task,
+  Notification,
 }