+++ ui/test/shared/components/MeasurementListItem.test.tsx
@@ -23,6 +23,7 @@ const setup = (overrides = {}) => {
     measurement: 'test',
     numTagsActive: 3,
     areTagsAccepted: true,
+    isQuerySupportedByExplorer: true,
     onChooseTag: () => {},
     onGroupByTag: () => {},
     onAcceptReject: () => {},